requirements.txt
flask>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=0.24.0
plotly>=5.0.0
scipy>=1.7.0
werkzeug>=2.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
catboost>=1.0.0
optuna>=2.10.0
shap>=0.40.0
imbalanced-learn>=0.8.0
joblib>=1.1.0

# Google Generative AI (Gemini)
google-generativeai>=0.3.0

# Business Intelligence & Analytics
dash>=2.0.0
dash-bootstrap-components>=1.0.0
dash-core-components>=2.0.0
dash-html-components>=2.0.0
dash-table>=5.0.0

# Email & Reporting
flask-mail>=0.9.1
reportlab>=3.6.0
openpyxl>=3.0.0
python-docx>=0.8.11
jinja2>=3.0.0

# Database & ORM
sqlalchemy>=1.4.0
flask-sqlalchemy>=2.5.0
alembic>=1.7.0
psycopg2-binary>=2.9.0

# Authentication & Security
flask-login>=0.5.0
flask-wtf>=1.0.0
werkzeug>=2.0.0
bcrypt>=3.2.0

# Background Tasks
celery>=5.2.0
redis>=4.0.0

# Time Series Analysis
statsmodels>=0.13.0
prophet>=1.1.0
pmdarima>=1.8.0

# Business Metrics & Analytics
scipy>=1.7.0
seaborn>=0.11.0
matplotlib>=3.5.0

# API & Web Services
requests>=2.25.0
gunicorn>=20.1.0

# Data Processing
python-dateutil>=2.8.0
pytz>=2021.1

# Configuration & Environment
python-dotenv>=0.19.0
configparser>=5.0.0

# Testing
pytest>=6.0.0
pytest-flask>=1.0.0

# Development
black>=21.0.0
flake8>=3.9.0
